refactoring code, added errrors, handlers

+++ src/main/java/se/foodload/application/FamilyService.java
@@ -7,6 +7,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import se.foodload.application.Interfaces.IFamilyService;
+import se.foodload.application.exception.ClientNotFoundException;
+import se.foodload.application.exception.FamilyInviteNotFoundException;
+import se.foodload.application.exception.FamilyNotFoundException;
 import se.foodload.domain.Client;
 import se.foodload.domain.Family;
 import se.foodload.domain.FamilyInvite;
@@ -40,11 +43,11 @@ public class FamilyService implements IFamilyService{
 	public Family changeFamilyName(long family_id, String newFamilyName) {
 		Optional<Family> foundFamily = familyRepo.findById(family_id);
 		if(foundFamily.isEmpty()) {
-			// THROW EXCEPTION.
+			throw new FamilyNotFoundException("No family with id "+family_id+" could be found.");
 		}
 		Family family = foundFamily.get();
 		family.setName(newFamilyName);
-		familyRepo.save(family); // ONÖDIG?
+		familyRepo.save(family); 
 		return family;
 	}
 
@@ -52,7 +55,7 @@ public class FamilyService implements IFamilyService{
 	public void inviteToFamily(Family family, String email) {
 		Optional<Client> client = clientRepo.findByEmail(email);
 		if(client.isEmpty()) {
-			//throw client not found with email x.
+			throw new ClientNotFoundException("No client could be found with email" + email);
 		}
 		FamilyInvite familyInv = new FamilyInvite(family, client.get()); 
 		familyInviteRepo.save(familyInv);	
@@ -64,7 +67,8 @@ public class FamilyService implements IFamilyService{
 	public FamilyInvite checkFamilyInvite(Client client) {
 		Optional<FamilyInvite> familyInv = familyInviteRepo.findByClientId(client);
 		if(familyInv.isEmpty()) {
-			//throw no invite error.
+			throw new FamilyInviteNotFoundException("No Family invite could be found for client "+ client.getEmail());
+			
 		}
 		return familyInv.get();
 	}
@@ -73,13 +77,19 @@ public class FamilyService implements IFamilyService{
 	public void acceptFamilyInvite(long familyInviteId) {
 		Optional<FamilyInvite> familyInvite =familyInviteRepo.findById(familyInviteId);
 		if(familyInvite.isEmpty()) {
-			//throw no invite error.
+			throw new FamilyInviteNotFoundException("No Family invite could be found with id "+ familyInviteId);
 		}
 		Client client = familyInvite.get().getClientId();
 		Family prevFamily = client.getFamily();
-		familyRepo.delete(prevFamily); //tar bort tidigare family, finns risk att man måste ta bort storages osv för sig.
+
+		Optional<List<Client>> clients = clientRepo.findByFamily(prevFamily);
+		if(clients.isEmpty()) {
+			familyRepo.delete(prevFamily); //tar bort tidigare family, finns risk att man måste ta bort storages osv för sig.
+		}
+		
 		client.addFamily(familyInvite.get().getFamilyId());
 		clientRepo.save(client);
+		familyInviteRepo.delete(familyInvite.get());
 		
 	}
